Separate source / implementation.
