+++ cpp/src/CMakeLists.txt
@@ -1,8 +1,10 @@
 add_library(toppra SHARED
+  toppra/piecewise_poly_path.cpp
   toppra/constraint.cpp
   toppra/constraint/linear_joint_velocity.cpp
   toppra/constraint/linear_joint_acceleration.cpp
   )
+
 target_include_directories(toppra PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
   $<INSTALL_INTERFACE:include>)
